feature: restructure component architecture so its easier to keep track of stuff
