--- src/components/Header.css
@@ -1,59 +1,0 @@
-.Header {
-    /* position: fixed; */
-    width: 100%;
-    padding: 50px 0;
-    z-index: 100; 
-}
-
-.HeaderGroup {
-    max-width: 800px;
-    margin: 0 auto;
-    display: grid;
-    grid-template-columns: repeat(5, auto);
-    align-items: center;
-    justify-items: center;
-}
-
-.Header a {
-    color: white;
-    font-weight: 700;
-}
-
-.Header button {
-    padding: 16px 20px;
-    font-size: 16px;
-    text-transform: uppercase;
-    color: white;
-    background: #EB493D;
-    border: none;
-    font-weight: 700;
-    border-radius: 8px;
-    outline: none;
-    cursor: pointer;
-    transition: 1s cubic-bezier(0.2, 0.8, 0.2, 1);
-}
-
-.Header button:hover {
-    background: white;
-    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.25);
-    transform: translateY(-3px);
-}
-
-@media (max-width: 640px) {
-    .Header {
-        padding: 15px 0;
-    }
-
-    .HeaderGroup {
-        grid-template-columns: repeat(4, auto);
-    }
-
-    .HeaderGroup a:nth-child(3) {
-        display: none;
-    }
-    
-    .HeaderGroup a:nth-child(4) {
-        display: none;
-    }
-}
-
