+++ src/components/Header/index.js
@@ -1,7 +1,5 @@
-import { Illustration } from './illustration'
-
+import { Illustration } from '../../components'
 import React from 'react'
-
 import styles from './styles.module.scss'
 
 const Header = ({ props }) => {
