+++ src/components/Illustration/styles.module.scss
