+++ src/components/index.js
@@ -1,0 +1,7 @@
+import Header from './Header'
+import Illustration from './Illustration'
+
+export {
+    Header,
+    Illustration,
+}
