+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styles from '../layouts/styles.module.scss'
-import { Header } from '../components/Header/index'
+import {Header} from '../components'
 
 const IndexPage = () => (
 	<div>
